+++ test/suite/test_layered_async_stepdown03.py
@@ -32,7 +32,7 @@ from helper_layered_stepdown import LayeredStepdownMixin
 from wtscenario import make_scenarios
 
 # test_layered_async_stepdown03.py
-#    Rollback guards: straddlers roll back; arm validation and timestamp guards.
+#    Straddler rollback guards: writers that cross the arm roll back.
 @disagg_test_class
 class test_layered_async_stepdown03(LayeredStepdownMixin, wttest.WiredTigerTestCase):
     conn_base_config = 'statistics=(all),statistics_log=(wait=1,json=true,on_close=true),'
@@ -203,7 +203,7 @@ class test_layered_async_stepdown03(LayeredStepdownMixin, wttest.WiredTigerTestC
         self.session.commit_transaction()
         rcur.close()
 
-    # Shared prefix: begin a stable write, arm, advance stable to the cutoff, checkpoint.
+    # Shared function: begin a stable write, arm, advance stable to the cutoff, checkpoint.
     def stable_writer_through_checkpoint(self):
         self.set_global_ts(1, 1)
         self.session.create(self.uri, 'key_format=S,value_format=S')
@@ -246,9 +246,7 @@ class test_layered_async_stepdown03(LayeredStepdownMixin, wttest.WiredTigerTestC
         self.session.begin_transaction()
         cursor['k1'] = 'v'
 
-        # The server completes the step-down under the in-flight ingest writer: stable advances
-        # to the cutoff, the step-down checkpoint is taken (on another session, as this one has
-        # the transaction open), the node reconfigures to follower.
+        # The server completes the step-down under the in-flight ingest writer.
         self.conn.set_timestamp('stable_timestamp=' + self.timestamp_str(20))
         ckpt_session = self.conn.open_session()
         ckpt_session.checkpoint()
@@ -260,98 +258,6 @@ class test_layered_async_stepdown03(LayeredStepdownMixin, wttest.WiredTigerTestC
         cursor.close()
         self.assertEqual(self.read_keys_at(self.ingest_uri(self.uri), 40), {'k1'})
         self.assertEqual(self.read_kvs_at(self.uri, 40), {'k1': 'v'})
-
-    # The cutoff cannot be re-armed while one is set.
-    def test_double_arm_rejected(self):
-        self.set_global_ts(1, 1)
-        self.session.create(self.uri, 'key_format=S,value_format=S')
-        self.arm(20)
-        self.assertRaisesWithMessage(wiredtiger.WiredTigerError,
-            lambda: self.arm(30), '/step down timestamp is already set/')
-
-    # Arming is only valid on a leader.
-    def test_arm_on_follower_rejected(self):
-        self.set_global_ts(1, 1)
-        self.session.create(self.uri, 'key_format=S,value_format=S')
-        self.conn.reconfigure('disaggregated=(role="follower")')
-        self.assertRaisesWithMessage(wiredtiger.WiredTigerError,
-            lambda: self.arm(20), '/can only be set on a disaggregated leader/')
-
-    # The cutoff must sit at or ahead of all_durable; arming exactly at it is allowed.
-    def test_arm_below_all_durable_rejected(self):
-        self.set_global_ts(1, 1)
-        self.session.create(self.uri, 'key_format=S,value_format=S')
-        self.write_at(self.uri, {'k1': 'v'}, 10)
-
-        self.assertRaisesWithMessage(wiredtiger.WiredTigerError,
-            lambda: self.arm(5), '/must not be older than all_durable timestamp/')
-        self.arm(10)
-
-    # While armed, stable may reach the cutoff exactly but never pass it.
-    def test_stable_cannot_pass_armed_cutoff(self):
-        self.set_global_ts(1, 1)
-        self.session.create(self.uri, 'key_format=S,value_format=S')
-        self.arm(20)
-
-        self.assertRaisesWithMessage(wiredtiger.WiredTigerError,
-            lambda: self.conn.set_timestamp('stable_timestamp=' + self.timestamp_str(25)),
-            '/must not advance past the armed step down timestamp/')
-        self.conn.set_timestamp('stable_timestamp=' + self.timestamp_str(20))
-
-    # Arming and advancing stable to the cutoff in one set_timestamp call takes full effect.
-    def test_arm_and_stable_in_one_call(self):
-        self.set_global_ts(1, 1)
-        self.session.create(self.uri, 'key_format=S,value_format=S')
-        self.conn.set_timestamp('stable_timestamp=' + self.timestamp_str(20) +
-                                ',step_down_timestamp=' + self.timestamp_str(20))
-
-        self.assertRaisesWithMessage(wiredtiger.WiredTigerError,
-            lambda: self.arm(30), '/step down timestamp is already set/')
-        self.assertRaisesWithMessage(wiredtiger.WiredTigerError,
-            lambda: self.conn.set_timestamp('stable_timestamp=' + self.timestamp_str(25)),
-            '/must not advance past the armed step down timestamp/')
-
-    # FIXME-WT-17895: one call that arms and overshoots stable is accepted; pins current behavior.
-    def test_arm_with_stable_overshoot_in_one_call(self):
-        self.set_global_ts(1, 1)
-        self.session.create(self.uri, 'key_format=S,value_format=S')
-        self.conn.set_timestamp('stable_timestamp=' + self.timestamp_str(25) +
-                                ',step_down_timestamp=' + self.timestamp_str(20))
-
-    # FIXME-WT-17895: arming below the current stable is accepted; pins current behavior.
-    def test_arm_below_stable_accepted(self):
-        self.set_global_ts(1, 10)
-        self.session.create(self.uri, 'key_format=S,value_format=S')
-        self.arm(5)
-
-    # all_durable is held below the cutoff by an in-flight pre-arm txn, passes it once resolved.
-    def test_all_durable_drain_signal_while_armed(self):
-        self.set_global_ts(1, 1)
-        self.session.create(self.uri, 'key_format=S,value_format=S')
-        self.write_at(self.uri, {'k1': 'v'}, 10)
-
-        # An in-flight transaction reserves commit timestamp 12; a later commit at 15 leaves a
-        # hole, so all_durable is clamped just below the reservation.
-        straddler = self.conn.open_session()
-        scur = straddler.open_cursor(self.uri, None, None)
-        straddler.begin_transaction()
-        scur['held'] = 'v'
-        straddler.timestamp_transaction('commit_timestamp=' + self.timestamp_str(12))
-        self.write_at(self.uri, {'k2': 'v'}, 15)
-        self.assertEqual(self.all_durable(), 11)
-
-        self.arm(20)
-        self.assertEqual(self.all_durable(), 11, 'arming must not move all_durable')
-
-        # The straddler resolves (the server rolls it back and retries): the hole closes.
-        straddler.rollback_transaction()
-        scur.close()
-        straddler.close()
-        self.assertEqual(self.all_durable(), 15)
-
-        # A post-arm commit above the cutoff carries all_durable past it: drained.
-        self.write_at(self.uri, {'k3': 'v'}, 25)
-        self.assertEqual(self.all_durable(), 25)
 
     # An untimestamped commit while armed succeeds and lands in ingest; pins current behavior.
     def test_untimestamped_commit_while_armed(self):
